Work on texture atlas

+++ ThreeDGame/src/particles/ParticleShader.java
@@ -34,10 +34,10 @@ public class ParticleShader extends ShaderProgram {
 		super.bindAttributes(0, "position");
 	}
 	
-	protected void loadTextureCoordInfo(Vector2f offset1, Vector2f offset2, float numRows, float blend){
+	protected void loadTextureCoordInfo(Vector2f offset1, Vector2f offset2, int numRows, float blend){
 		super.load2DVector(location_texOffset1, offset1);
 		super.load2DVector(location_texOffset2, offset2);
-		super.load2DVector(location_texCoordInfo, new Vector2f(numRows, blend));
+		super.load2DVector(location_texCoordInfo, new Vector2f((float)numRows, blend));
 	}
 
 	protected void loadModelViewMatrix(Matrix4f modelView) {
